Rename com.vaadin.data.util.converter to com.vaadin.data.converter
Part of vaadin/framework8-issues#110
