+++ server/src/main/java/com/vaadin/data/converter/DateToSqlDateConverter.java
@@ -17,7 +17,7 @@
 /**
  *
  */
-package com.vaadin.data.util.converter;
+package com.vaadin.data.converter;
 
 import java.util.Date;
