+++ server/src/main/java/com/vaadin/data/converter/StringToDateConverter.java
@@ -14,7 +14,7 @@
  * the License.
  */
 
-package com.vaadin.data.util.converter;
+package com.vaadin.data.converter;
 
 import java.text.DateFormat;
 import java.text.ParsePosition;
