+++ server/src/main/java/com/vaadin/data/converter/ValueContext.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.data.util.converter;
+package com.vaadin.data.converter;
 
 import java.io.Serializable;
 import java.util.Locale;
