+++ server/src/main/java/com/vaadin/data/validator/BeanValidator.java
@@ -30,8 +30,8 @@ import javax.validation.metadata.ConstraintDescriptor;
 
 import com.vaadin.data.ValidationResult;
 import com.vaadin.data.Validator;
+import com.vaadin.data.converter.ValueContext;
 import com.vaadin.data.util.BeanUtil;
-import com.vaadin.data.util.converter.ValueContext;
 
 /**
  * A {@code Validator} using the JSR-303 (javax.validation) annotation-based
