+++ server/src/main/java/com/vaadin/ui/declarative/converters/DesignShortcutActionConverter.java
@@ -21,8 +21,8 @@ import java.util.Map;
 import java.util.Map.Entry;
 
 import com.vaadin.data.Result;
-import com.vaadin.data.util.converter.Converter;
-import com.vaadin.data.util.converter.ValueContext;
+import com.vaadin.data.converter.Converter;
+import com.vaadin.data.converter.ValueContext;
 import com.vaadin.event.ShortcutAction;
 import com.vaadin.event.ShortcutAction.KeyCode;
 import com.vaadin.event.ShortcutAction.ModifierKey;
