+++ server/src/main/java/com/vaadin/ui/declarative/converters/DesignTimeZoneConverter.java
@@ -18,8 +18,8 @@ package com.vaadin.ui.declarative.converters;
 import java.util.TimeZone;
 
 import com.vaadin.data.Result;
-import com.vaadin.data.util.converter.Converter;
-import com.vaadin.data.util.converter.ValueContext;
+import com.vaadin.data.converter.Converter;
+import com.vaadin.data.converter.ValueContext;
 import com.vaadin.ui.declarative.DesignAttributeHandler;
 
 /**
