+++ server/src/test/java/com/vaadin/data/BinderTest.java
@@ -14,7 +14,7 @@ import org.junit.Before;
 import org.junit.Test;
 
 import com.vaadin.data.Binder.BindingBuilder;
-import com.vaadin.data.util.converter.StringToIntegerConverter;
+import com.vaadin.data.converter.StringToIntegerConverter;
 import com.vaadin.data.validator.NotEmptyValidator;
 import com.vaadin.server.ErrorMessage;
 import com.vaadin.tests.data.bean.Person;
