+++ server/src/test/java/com/vaadin/data/validator/NotEmptyValidator.java
@@ -18,10 +18,10 @@ package com.vaadin.data.validator;
 import java.util.Objects;
 
 import com.vaadin.data.Binder.BindingBuilder;
+import com.vaadin.data.converter.ValueContext;
 import com.vaadin.data.HasValue;
 import com.vaadin.data.ValidationResult;
 import com.vaadin.data.Validator;
-import com.vaadin.data.util.converter.ValueContext;
 
 /**
  * Simple validator to check against {@code null} value and empty {@link String}
