+++ server/src/test/java/com/vaadin/data/validator/NotEmptyValidatorTest.java
@@ -19,7 +19,7 @@ import org.junit.Assert;
 import org.junit.Test;
 
 import com.vaadin.data.ValidationResult;
-import com.vaadin.data.util.converter.ValueContext;
+import com.vaadin.data.converter.ValueContext;
 
 /**
  * @author Vaadin Ltd
