+++ server/src/test/java/com/vaadin/tests/data/converter/AbstractStringConverterTest.java
@@ -4,8 +4,8 @@ import org.junit.Assert;
 import org.junit.Test;
 
 import com.vaadin.data.Result;
-import com.vaadin.data.util.converter.Converter;
-import com.vaadin.data.util.converter.ValueContext;
+import com.vaadin.data.converter.Converter;
+import com.vaadin.data.converter.ValueContext;
 
 public abstract class AbstractStringConverterTest
         extends AbstractConverterTest {
