+++ server/src/test/java/com/vaadin/tests/data/converter/StringToBigIntegerConverterTest.java
@@ -22,8 +22,8 @@ import org.junit.Assert;
 import org.junit.Test;
 
 import com.vaadin.data.Result;
-import com.vaadin.data.util.converter.StringToBigIntegerConverter;
-import com.vaadin.data.util.converter.ValueContext;
+import com.vaadin.data.converter.StringToBigIntegerConverter;
+import com.vaadin.data.converter.ValueContext;
 
 public class StringToBigIntegerConverterTest
         extends AbstractStringConverterTest {
