+++ server/src/test/java/com/vaadin/tests/data/converter/StringToDoubleConverterTest.java
@@ -4,8 +4,8 @@ import org.junit.Assert;
 import org.junit.Test;
 
 import com.vaadin.data.Result;
-import com.vaadin.data.util.converter.StringToDoubleConverter;
-import com.vaadin.data.util.converter.ValueContext;
+import com.vaadin.data.converter.StringToDoubleConverter;
+import com.vaadin.data.converter.ValueContext;
 
 public class StringToDoubleConverterTest extends AbstractConverterTest {
