+++ server/src/test/java/com/vaadin/tests/data/converter/StringToLongConverterTest.java
@@ -3,8 +3,8 @@ package com.vaadin.tests.data.converter;
 import org.junit.Test;
 
 import com.vaadin.data.Result;
-import com.vaadin.data.util.converter.StringToLongConverter;
-import com.vaadin.data.util.converter.ValueContext;
+import com.vaadin.data.converter.StringToLongConverter;
+import com.vaadin.data.converter.ValueContext;
 
 public class StringToLongConverterTest extends AbstractStringConverterTest {
